# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Job Tracker Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List all companies for current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    patch:
      operationId: updateCompany
      tags: [companies]
      summary: Update a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    delete:
      operationId: deleteCompany
      tags: [companies]
      summary: Delete a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Job Applications
...
  /payments/subscription:
    get:
      operationId: getSubscription
      tags: [payments]
      summary: Get current user subscription
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
    post:
      operationId: updateSubscription
      tags: [payments]
      summary: Upgrade or downgrade plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /ai/analyze/{applicationId}:
    get:
      operationId: analyzeApplication
      tags: [ai]
      summary: Analyze a specific application and return AI suggestions
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAnalysis"
components:
  schemas:
...
    Company:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        website:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        applicationCount:
          type: integer
        createdAt:
          type: string
    CompanyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        website:
          type: string
        industry:
          type: string
        location:
          type: string
        notes:
          type: string
    CompanyUpdate:
      type: object
      properties:
        name:
          type: string
        website:
          type: string
        industry:
          type: string
        location:
          type: string
        notes:
          type: string

    # Applications
...
    Subscription:
      type: object
      required: [id, planId, planName, status, currentPeriodEnd]
      properties:
        id:
          type: integer
        planId:
          type: string
        planName:
          type: string
        status:
          type: string
        currentPeriodEnd:
          type: string
        cancelAtPeriodEnd:
          type: boolean
    SubscriptionInput:
      type: object
      required: [planId]
      properties:
        planId:
          type: string
...
    Notification:
      type: object
      required: [id, type, title, message, read, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        applicationId:
          type: ["integer", "null"]
        createdAt:
          type: string

    # AI
...
    AiAnalysis:
      type: object
      required: [applicationId, score, suggestions, nextSteps]
      properties:
        applicationId:
          type: integer
        score:
          type: integer
        suggestions:
          type: array
          items:
            type: string
        nextSteps:
          type: array
          items:
            type: string
        estimatedResponse:
          type: ["string", "null"]